pyproject.toml
[tool.poetry]
name = "tradeforce"
version = "0.2.0"
description = "Framework for Backtesting and Live-Trading. Hyperparameter-Optimization of Trading-Strategies through performant simulations."
authors = ["cyclux <nevermind@cyclux.de>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
numpy = "<1.24"
numba = "^0.56.4"
pandas = "^1.5.3"
pyarrow = "^11.0.0"
optuna = "^3.1.0"
tqdm = "^4.65.0"
pymongo = "^4.3.3"
psycopg2-binary = "^2.9.5"
pyyaml = "^6.0"
asyncio = "^3.4.3"
tensorflow-probability = "^0.19.0"
numexpr = { version = "^2.8.4", optional = true }
bottleneck = { version = "^1.3.7", optional = true }
bitfinex-api-py = { git = "https://github.com/cyclux/bitfinex-api-py.git" }
cryptography = "^40.0.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.260"
mypy = "^1.1.1"
types-psycopg2 = "^2.9.21.9"
types-pytz = "^2023.2.0.1"
types-pyyaml = "^6.0.12.9"
types-tqdm = "^4.65.0.1"
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
pandas-stubs = "^1.5.3.230321"
ipykernel = "^6.22.0"
ipywidgets = "^8.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.mypy]
exclude = ['archive', 'analysis']

[tool.pytest.ini_options]
testpaths = ['tests']
